move views in styleguide into module
